+++ packages/workload-router/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2026-05-06
+
+### Released
+
+- v0.6.0
+
+## [0.5.4] - 2026-05-02
+
+### Dependencies
+
+- Sync package versions to 0.5.4
+
 ## [0.5.1] - 2026-04-30
 
 ### Added
@@ -24,4 +36,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Released
 
 - v0.4.1
-